enum to float deprecated fix

+++ writeengine/server/we_dataloader.cpp
@@ -901,7 +901,7 @@ void WEDataLoader::onReceiveData(ByteStream& Ibs)
             // Bug 5031 : If Q size goes above 100 (2*250);
             if (aQsz < MAX_QSIZE) sendDataRequest();
 
-            if (aQsz > 1.5 * MAX_QSIZE) // > 2*250
+            if (aQsz > 1.5 * static_cast<int>(MAX_QSIZE)) // > 2*250
             {
                 cout << "WARNING : Data Queuing up : QSize = " << aQsz << endl;
